Add image rotation; prep enhancements

+++ bin/op-prep.py
@@ -132,33 +132,39 @@ def clean_dir(source_dir, clobber_pattern):
                 path = os.path.join(root, name)
                 os.remove(path)
 
-def redo_document(doc):
+def redo_document(doc, opts):
     if logger.getEffectiveLevel() >= logging.INFO:
         msg = "Preparing to redo document id: %d,"
         msg += " repository: %s, base_dir: %s"
         logger.info(msg % doc.id, doc.collection, doc.base_dir)
-    if doc.is_online and str(os.getenv('OPENN_REDO_OK_IF_YOU_SAY_SO', None)).lower() != 'true':
+    if doc.is_online and str(os.getenv('OPENN_REDO_OVERRIDE_ONLINE_HALT', None)).lower() != 'true':
         msg = "Redo requested, but refusing to redo record "
         msg += "for document on-line at: %s" % (doc.package_dir,)
         raise OPennException(msg)
     else:
-        s=  None
-        while s is None:
-            s = raw_input("Proceed with redo? Type Yes or No: ")
-            s = s.strip()
-            if s is not None:
-                if s.lower() == 'yes':
-                    logger.info("OK. Removing images from existing document.")
-                    doc.image_set.all().delete()
-                elif s.lower() == 'no':
-                    msg = "User canceled redo; no changes made"
-                    raise OPennException(msg)
-                else:
-                    msg = "Please enter 'Yes' or 'No'. I don't understand: %r"
-                    print msg % (s,)
-                    s = None
+        if opts.yes is True:
+            pass
+        elif opts.no is True:
+            msg = "User canceled redo; no changes made"
+            raise OPennException(msg)
+        else:
+            s=  None
+            while s is None:
+                s = raw_input("Proceed with redo? Type Yes or No: ")
+                s = s.strip()
+                if s is not None:
+                    if s.lower() == 'yes':
+                        logger.info("OK. Removing images from existing document.")
+                        doc.image_set.all().delete()
+                    elif s.lower() == 'no':
+                        msg = "User canceled redo; no changes made"
+                        raise OPennException(msg)
+                    else:
+                        msg = "Please enter 'Yes' or 'No'. I don't understand: %r"
+                        print msg % (s,)
+                        s = None
 
-def clobber_document(params):
+def clobber_document(params, opts):
     doc = openn_db.get_doc(params)
 
     if logger.getEffectiveLevel() >= logging.INFO:
@@ -171,21 +177,27 @@ def clobber_document(params):
         msg += "for document on-line at: %s" % (doc.package_dir,)
         raise OPennException(msg)
     else:
-        s = None
-        while s is None:
-            s = raw_input("Proceed with clobber? Type Yes or No: ")
-            s = s.strip()
-            if s is not None:
-                if s.lower() == 'yes':
-                    logger.info("OK. Deleting existing document.")
-                    doc.delete()
-                elif s.lower() == 'no':
-                    msg = "User canceled clobber; no changes made"
-                    raise OPennException(msg)
-                else:
-                    msg = "Please enter 'Yes' or 'No'. I don't understand: %r"
-                    print msg % (s,)
-                    s = None
+        if opts.yes is True:
+            pass
+        elif opts.no is True:
+            msg = "User canceled clobber; no changes made"
+            raise OPennException(msg)
+        else:
+            s = None
+            while s is None:
+                s = raw_input("Proceed with clobber? Type Yes or No: ")
+                s = s.strip()
+                if s is not None:
+                    if s.lower() == 'yes':
+                        logger.info("OK. Deleting existing document.")
+                        doc.delete()
+                    elif s.lower() == 'no':
+                        msg = "User canceled clobber; no changes made"
+                        raise OPennException(msg)
+                    else:
+                        msg = "Please enter 'Yes' or 'No'. I don't understand: %r"
+                        print msg % (s,)
+                        s = None
 
 def prep_source_dir_arg(source_dir):
     if source_dir.strip().endswith('/'):
@@ -265,7 +277,7 @@ def main(cmdline=None):
         if opts.clobber:
             if openn_db.doc_exists(doc_params):
                 try:
-                    clobber_document(doc_params)
+                    clobber_document(doc_params, opts)
                 except OPennException as ex:
                     parser.error(str(ex))
             else:
@@ -276,7 +288,7 @@ def main(cmdline=None):
             if openn_db.doc_exists(doc_params):
                 try:
                     doc = openn_db.get_doc(doc_params)
-                    redo_document(doc)
+                    redo_document(doc, opts)
                 except OPennException as ex:
                     parser.error(str(ex))
             else:
@@ -352,6 +364,13 @@ prepare correctly the first time. Will fail if document is on-line.
     parser.add_option('-s', '--soft-clobber',
                      action='store_true', dest='redo', default=False,
                      help=redo_help)
+
+    parser.add_option('-y', '--yes', action='store_true', dest='yes', default=False,
+                    help="Reply 'yes' to prompts")
+
+    parser.add_option('-n', '--no', action='store_true', dest='no', default=False,
+                    help="Reply 'no' to prompts")
+
 
     return parser
